refactor test and remove todo in comments
