+++ test/legacy_test/test_fused_multi_transformer_op.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
 import random
 import unittest
 
@@ -22,7 +21,7 @@ from test_sparse_attention_op import get_cuda_version
 
 import paddle
 import paddle.nn.functional as F
-from paddle import tensor
+from paddle import base, tensor
 from paddle.base.framework import default_main_program
 from paddle.base.param_attr import ParamAttr
 from paddle.incubate.nn import FusedMultiTransformer
@@ -1713,7 +1712,6 @@ class TestFusedMultiTransformerOpVariableDecoder2(TestFusedMultiTransformerOp):
         self.x_type = np.float16
 
 
-
 @unittest.skipIf(
     not paddle.is_compiled_with_cuda()
     or get_cuda_version() < 11030
@@ -1834,6 +1832,30 @@ class TestFusedMultiTransformerOpVariableGQAGenCache3(
         self.layers = 4  # even layers
         self.rotary_emb_dims = 2
         self.x_type = np.float16
+
+
+@unittest.skipIf(
+    not paddle.is_compiled_with_cuda()
+    or get_cuda_version() < 11030
+    or paddle.device.cuda.get_device_capability()[0] < 8,
+    "FusedMultiTransformer requires CUDA >= 11.2 and CUDA_ARCH >= 8",
+)
+class TestFusedMultiTransformerOpUseMBMMHA(TestFusedMultiTransformerOp):
+    def config(self):
+        super().config()
+        base.set_flags({'FLAGS_fused_multi_transformer_op_use_mbfmha': True})
+        self.has_cache_kv = True
+        self.gen_cache_kv = False
+        self.remove_padding = True
+        self.query_length = 1
+        self.key_length, self.value_length = 1, 1
+        self.cache_length = 2049
+        # NOTE(Wanglongzhi2001): cache length is little long, to avoid the cost time of this unittest is too long,
+        # we set the layers to 2.
+        self.layers = 2
+        self.rotary_emb_dims = 2
+        self.x_type = np.float16
+
 
 if __name__ == "__main__":
     unittest.main()
